Video-Download-to-NAS/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
yt-dlp==2024.3.10
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
websockets==12.0
slowapi==0.1.9

# SSO Authentication dependencies
httpx==0.25.2
authlib==1.3.0
cryptography==42.0.0
apscheduler==3.10.4

# Telegram Bot dependencies
python-telegram-bot==20.7
aiohttp==3.9.1

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.0
